Fix warnings reported by -Wextra
- use unsigned loop indices where they only index GPtrArray/GArray
  and cast explicitly where signed API values are compared with unsigned
  ones
- make LspUtilsCmpFn and lsp_utils_lowercase_cmp() return gint instead
  of passing the result through gpointer so callers don't need function
  pointer casts
- make get_ident_prefixlen() return gint which is what it computes
- initialize both positions of LspRange explicitly
- avoid unsigned arithmetic in the MAX() expression of
  lsp_progress_free_all()

Unused callback parameters are left as they are.

+++ lsp/src/lsp-main.c
@@ -1288,7 +1288,7 @@ static gboolean on_code_actions_received_kb(GPtrArray *code_action_commands, gpo
 
 		foreach_ptr_array(cmd, i, code_lens_commands)
 		{
-			if (cmd->line == line &&
+			if ((gint)cmd->line == line &&
 				g_regex_match_simple(cmd_str, cmd->title, G_REGEX_CASELESS, G_REGEX_MATCH_NOTEMPTY))
 			{
 				lsp_command_perform(srv, cmd, NULL, NULL);
